refactor(dataconverter): split samples into per-pattern packages
