+++ src/main/java/com/uber/cadence/samples/compression/CompressedJsonDataConverter.java
@@ -15,7 +15,7 @@
  *  permissions and limitations under the License.
  */
 
-package com.uber.cadence.samples.dataconverter;
+package com.uber.cadence.samples.compression;
 
 import com.uber.cadence.converter.DataConverter;
 import com.uber.cadence.converter.DataConverterException;
