+++ src/main/java/com/uber/cadence/samples/encryption/EncryptionKeyLoader.java
@@ -15,7 +15,7 @@
  *  permissions and limitations under the License.
  */
 
-package com.uber.cadence.samples.dataconverter;
+package com.uber.cadence.samples.encryption;
 
 import java.nio.charset.StandardCharsets;
