+++ src/main/java/com/uber/cadence/samples/s3offload/S3OffloadDataConverter.java
@@ -15,7 +15,7 @@
  *  permissions and limitations under the License.
  */
 
-package com.uber.cadence.samples.dataconverter;
+package com.uber.cadence.samples.s3offload;
 
 import com.uber.cadence.converter.DataConverter;
 import com.uber.cadence.converter.DataConverterException;
@@ -79,7 +79,7 @@ import java.security.NoSuchAlgorithmException;
  * }
  *
  *  3. Replace `new LocalFsBlobStore()` with `new S3BlobStore("my-bucket", "us-east-1")` in
- *     DataConverterWorker.
+ *     S3OffloadWorker.
  *  4. Set standard AWS env vars (AWS_REGION, AWS_ACCESS_KEY_ID, AWS_SECRET_ACCESS_KEY) or use an
  *     IAM instance role.
  *
